split total_xsection into the different parmaterizations for ease of expanding
